cli: overhaul and separate advanced LoRA commands
This should make our CLI much more discoverable.
