+++ backend/flask.sh
@@ -11,13 +11,11 @@ then
 
     "$VENV"/bin/pip install -qU packaging pip setuptools wheel
     "$VENV"/bin/pip install -qr "$DIR"/requirements-test.txt
-    "$VENV"/bin/pip install -qe "$DIR"
 fi
 
-# detected by 'cli.py', gets us the appropriate "Usage:" printout
-export MORA_PROG_NAME="$0"
+"$VENV"/bin/pip install -qe "$DIR"
 
-# we also have '$VENV/bin/python', but use python -m so that we can
-# override the program name
+export FLASK_ENV=development
+export FLASK_APP=mora.app:app
 
-exec "$VENV"/bin/flask "$@"
+exec "$VENV"/bin/python -m mora.cli "$0" "$@"
